Add support to '%u' format specifier, updating 'write_number' function to change the data type of input parameter from 'int/unsigned int' to 'long/unsigned long' and updating nested if..then's of 'ft_printf.c' to add '%u'

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: jfidalgo <jfidalgo@student.42bar(...).com  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/22 15:19:42 by jfidalgo          #+#    #+#             */
-/*   Updated: 2024/02/26 17:53:06 by jfidalgo         ###   ########.fr       */
+/*   Updated: 2024/02/28 12:20:41 by jfidalgo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 
 int	write_char(int character);
 int	write_string(char *string);
-int	write_number(int number);
+long	write_number(long number);
 int	write_hexa(int number, int uppercase);
 int	ft_printf(char const *format, ...);
